Renamed BraintreeGateway
